refactor: move ui parts from web to shared
