+++ projects/lp/src/app/pages/home/home.component.ts
@@ -1,11 +1,11 @@
-import { FaArrowRightFromBracketIconComponent } from '@/web/src/app/ui/icons/fontawesome/fa-arrow-right-from-bracket-icon/fa-arrow-right-from-bracket-icon.component';
-import { FaArrowRightToBracketIconComponent } from '@/web/src/app/ui/icons/fontawesome/fa-arrow-right-to-bracket-icon/fa-arrow-right-to-bracket-icon.component';
-import { FaBarsIconComponent } from '@/web/src/app/ui/icons/fontawesome/fa-bars-icon/fa-bars-icon.component';
-import { FaGoogleIconComponent } from '@/web/src/app/ui/icons/fontawesome/fa-google/fa-google-icon.component';
-import { FaHomeIconComponent } from '@/web/src/app/ui/icons/fontawesome/fa-home-icon/fa-home-icon.component';
-import { FaQrcodeIconComponent } from '@/web/src/app/ui/icons/fontawesome/fa-qrcode-icon/fa-qrcode-icon.component';
-import { FaUserIconComponent } from '@/web/src/app/ui/icons/fontawesome/fa-user-icon/fa-user-icon.component';
-import { FaUserPlusIconComponent } from '@/web/src/app/ui/icons/fontawesome/fa-user-plus-icon/fa-user-plus-icon.component';
+import { FaArrowRightFromBracketIconComponent } from '@/shared/src/app/ui/icons/fontawesome/fa-arrow-right-from-bracket-icon/fa-arrow-right-from-bracket-icon.component';
+import { FaArrowRightToBracketIconComponent } from '@/shared/src/app/ui/icons/fontawesome/fa-arrow-right-to-bracket-icon/fa-arrow-right-to-bracket-icon.component';
+import { FaBarsIconComponent } from '@/shared/src/app/ui/icons/fontawesome/fa-bars-icon/fa-bars-icon.component';
+import { FaGoogleIconComponent } from '@/shared/src/app/ui/icons/fontawesome/fa-google/fa-google-icon.component';
+import { FaHomeIconComponent } from '@/shared/src/app/ui/icons/fontawesome/fa-home-icon/fa-home-icon.component';
+import { FaQrcodeIconComponent } from '@/shared/src/app/ui/icons/fontawesome/fa-qrcode-icon/fa-qrcode-icon.component';
+import { FaUserIconComponent } from '@/shared/src/app/ui/icons/fontawesome/fa-user-icon/fa-user-icon.component';
+import { FaUserPlusIconComponent } from '@/shared/src/app/ui/icons/fontawesome/fa-user-plus-icon/fa-user-plus-icon.component';
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
@@ -27,14 +27,14 @@ import { Component } from '@angular/core';
     <div class="prose">
       <h1>Hello world!</h1>
       <p>
-        <web-fa-arrow-right-from-bracket-icon></web-fa-arrow-right-from-bracket-icon>
-        <web-fa-arrow-right-to-bracket-icon></web-fa-arrow-right-to-bracket-icon>
-        <web-fa-bars-icon></web-fa-bars-icon>
-        <web-fa-google-icon></web-fa-google-icon>
-        <web-fa-home-icon></web-fa-home-icon>
-        <web-fa-qrcode-icon></web-fa-qrcode-icon>
-        <web-fa-user-icon></web-fa-user-icon>
-        <web-fa-user-plus-icon></web-fa-user-plus-icon>
+        <shared-fa-arrow-right-from-bracket-icon></shared-fa-arrow-right-from-bracket-icon>
+        <shared-fa-arrow-right-to-bracket-icon></shared-fa-arrow-right-to-bracket-icon>
+        <shared-fa-bars-icon></shared-fa-bars-icon>
+        <shared-fa-google-icon></shared-fa-google-icon>
+        <shared-fa-home-icon></shared-fa-home-icon>
+        <shared-fa-qrcode-icon></shared-fa-qrcode-icon>
+        <shared-fa-user-icon></shared-fa-user-icon>
+        <shared-fa-user-plus-icon></shared-fa-user-plus-icon>
       </p>
       <h2>H2 Test</h2>
       <p>
