+++ projects/shared/src/app/ui/icons/fontawesome/fa-arrow-right-from-bracket-icon/fa-arrow-right-from-bracket-icon.component.spec.ts
