+++ projects/shared/src/app/ui/icons/fontawesome/fa-google/fa-google-icon.component.ts
@@ -1,15 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { faBars } from '@fortawesome/free-solid-svg-icons';
+import { faGoogle } from '@fortawesome/free-brands-svg-icons';
 
 @Component({
-  selector: 'web-fa-bars-icon',
+  selector: 'shared-fa-google-icon',
   standalone: true,
   imports: [CommonModule, FontAwesomeModule],
   template: ` <fa-icon [icon]="iconDefinition"></fa-icon> `,
   styles: [],
 })
-export class FaBarsIconComponent {
-  iconDefinition = faBars;
+export class FaGoogleIconComponent {
+  iconDefinition = faGoogle;
 }
