+++ projects/shared/src/app/ui/icons/fontawesome/fa-qrcode-icon/fa-qrcode-icon.component.ts
@@ -1,15 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { faArrowRightFromBracket } from '@fortawesome/free-solid-svg-icons';
+import { faQrcode } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
-  selector: 'web-fa-arrow-right-from-bracket-icon',
+  selector: 'shared-fa-qrcode-icon',
   standalone: true,
   imports: [CommonModule, FontAwesomeModule],
   template: ` <fa-icon [icon]="iconDefinition"></fa-icon> `,
   styles: [],
 })
-export class FaArrowRightFromBracketIconComponent {
-  iconDefinition = faArrowRightFromBracket;
+export class FaQrcodeIconComponent {
+  iconDefinition = faQrcode;
 }
