+++ projects/shared/src/app/ui/icons/fontawesome/fa-user-icon/fa-user-icon.component.ts
@@ -1,15 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { faQrcode } from '@fortawesome/free-solid-svg-icons';
+import { faUser } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
-  selector: 'web-fa-qrcode-icon',
+  selector: 'shared-fa-user-icon',
   standalone: true,
   imports: [CommonModule, FontAwesomeModule],
   template: ` <fa-icon [icon]="iconDefinition"></fa-icon> `,
   styles: [],
 })
-export class FaQrcodeIconComponent {
-  iconDefinition = faQrcode;
+export class FaUserIconComponent {
+  iconDefinition = faUser;
 }
