+++ projects/shared/src/app/ui/icons/fontawesome/fa-user-plus-icon/fa-user-plus-icon.component.ts
@@ -1,15 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { faHome } from '@fortawesome/free-solid-svg-icons';
+import { faUserPlus } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
-  selector: 'web-fa-home-icon',
+  selector: 'shared-fa-user-plus-icon',
   standalone: true,
   imports: [CommonModule, FontAwesomeModule],
   template: ` <fa-icon [icon]="iconDefinition"></fa-icon> `,
   styles: [],
 })
-export class FaHomeIconComponent {
-  iconDefinition = faHome;
+export class FaUserPlusIconComponent {
+  iconDefinition = faUserPlus;
 }
